mz-deploy: pull did_you_mean into a shared suggest module
The LSP's quick fixes rank candidate names by Damerau-Levenshtein
distance to suggest a replacement for a name that did not resolve. The
deployment validation errors want the same ranking, and having `client`
reach into `lsp` for it would be the wrong dependency.

Moves the helper and its tests to `suggest`, unchanged, so a misspelling
can read the same however the user hit it.
